Revert "Add hero transition for countdown cards"
