+++ Services/Haptics.swift
@@ -4,9 +4,6 @@ enum Haptics {
     static func light() {
         UIImpactFeedbackGenerator(style: .light).impactOccurred()
     }
-    static func medium() {
-        UIImpactFeedbackGenerator(style: .medium).impactOccurred()
-    }
     static func success() {
         UINotificationFeedbackGenerator().notificationOccurred(.success)
     }
